Send GitHub HipChat messages ourselves
Summary:
With this, you can add https://khan-webhooks.appspot.com/github-feed as a custom web hook and we'll get HipChat notifications. This allows us to send branch notifications to other rooms, which I've set up here for the "athena" branch.

This should match the existing message format almost exactly. I took the liberty of adding a line "- and 3 existing commits" when fast-forward pushing a branch, which I think is helpful.

Test Plan: Deployed it with it set to send to "HipChat Tests", had GitHub send test payloads (at https://github.com/Khan/webapp/settings/hooks/2551047), and observed that the messages looked approximately the same as before.

Reviewers: marcos, benkomalo

Reviewed By: benkomalo

Subscribers: eater

Differential Revision: http://phabricator.khanacademy.org/D10801