Clean up extra PagerParrot messages
Summary:
I forgot to have it only send to slack on the "trigger" message, so it sent a
bunch of pings when we acknowledged and resolved the alerts.  This should no
longer happen.  In addition, PagerDuty sent a few duplicate POSTS, which they
say they might do.  So we'll use the UUID they provide to deduplicate.

If we find ourselves actually wanting more parrotyness, we can always just have
it send twice for each message it gets :-) .

Test Plan: Will do another full test of p0@ after deploying this.

Reviewers: benjaminpollack

Subscribers: csilvers, alpert

Differential Revision: https://phabricator.khanacademy.org/D24029