Add message summary to Pager Parrot alerts
People can click through to see it, but then they have to look up the
credentials, type them in, etc, which is confusing and annoying when you're not
sure if the site is down.  Now we'll put in the summary if we have it.

Test plan: deploy, and trigger a test P0.

Auditors: benjaminpollack.