Fix Downtime Alert Summary Messages
Summary:
Previously, downtime alerts triggered by our Google monitoring would not have a summary in Slack, and responders would need to click in to the PagerDuty incident to get details of what was actually going on.

(Previously)
{F1442139}

The problem was no short summary is provided from the Google integration. Instead a full-length summary is used for the problem.

This change uses the short one if we have it, and falls back to the longer one if we don't.

To account for the longer message, the text was reformatted to put the summary on a new line. Hopefully this will also make it easier to see what's going on at a glance.

(Now)
{F1442140}

(With a short summary)
{F1442142}

Test Plan: Pulled several actual incident messages from the logs and tried them out

Reviewers: amos, alexandershah

Reviewed By: amos

Subscribers: csilvers, boris

Differential Revision: https://phabricator.khanacademy.org/D54841