Add logging statements to new PhabFox endpoint
Summary: In switching to the new PhabFox system outlined in INFRA-2745, a couple initial logging statements are being added for when the new endpoint `new-phabricator-feed` is triggered to see what the new object format is (and what data is available).

Test Plan: Manually checking things. Working on this new webhook in parallel with keeping the old one, so nothing should be affected, but will keep an eye out for anything weird.

Reviewers: jacqueline

Reviewed By: jacqueline

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D55930