Check null transactions case and use author username
Summary:
`transaction.search` requires a non-empty list as constraints, so I added a check to see if the parameter is an empty list before making the function call. Also, the for loop on line 244 requires the author's username to go through the user_channel_map, so I added a function that gets the author's username based on the authorPHID.

In addition, to decrease the amount of nesting and misleading-ness caused by the `else` being so far from its `if` counterpart, I restructured the code to follow `if not ...` pattern and to return out of the post() function early if there is no response or no transactions.

Errors from logs:
{F1468353}
{F1468354}

Test Plan: Some more manual testing through logs on GCP

Reviewers: jacqueline, csilvers

Reviewed By: jacqueline, csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D56195