Thread Multiple Pager Parrot Alerts in a Row
Summary:
If Pager Parrot is posting a lot messages quickly, it will post them into a thread, rather than completely take over a channel with alerts.

Most commonly this channel spam will happen for downtime alerts during a major outage, and it can make it difficult to have discussions about how to address the problem in the channel.

Now it will be compacted and only stop posting to a thread after a 15 minute break between alerts. This is less than the @channel threshold, but should make sure like alerts are grouped without hiding alerts we need to be aware of.

To do the threading, I've had to switch from using a webhook url to a bot authentication token, and that will need to be accounted for at deploy time.

Deploy Plan:
Create new slack bot
Invite bot to all the correct channels (`1s-and-0s`, `content`, and `user_issues`)
Save bot token in Keeper
Deploy live
Let channels know that an alert will be coming and it is just a test
Fire test alert

Test Plan:
Run and trigger alerts one after another
Wait 15 minutes and trigger another alert

Reviewers: amos, boris

Reviewed By: amos

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D57863