Upgrading PagerDuty to webhook v3
PagerDuty is deprecating it's legacy v1 webhook.  So we are adapting to
the v3 webhook.

This is relatively straightforward in adapting the Webhook.  But now
looking at the remaining usage of the webhook with Phabricator (already
deprecated) - it may be useful to move away from using khan-webhook for
notifying slack.

Issue: https://khanacademy.atlassian.net/browse/INFRA-8511

Test plan:
- Create test incident with our test project, and ensure the integration
  is working.